working on fixes for hexes and map loading

+++ src/database/queries/getHexDetails.sql
@@ -1,3 +1,3 @@
 select *
 from kingmaker."hex"
-where x_coord = $1 and y_coord = $2 and map_id = $3
+where x_coord = $1::integer and y_coord = $2::integer and map_id = $3::integer
